Use FAS to tell if GitHub user is a Fedora packager
Querying badges.fedoraproject.org to tell if a user is a packager in
Fedora is slow: takes seconds for non existing usernames and can take
over a minute for an existing one.

Query FAS instead, and check if the users GitHub username is in the
'packager' group. This is faster and probably also more correct, as FAS
is the central place to manage users in Fedora space.

Add a new property (`fas_password`) to the service configuration to be
used to authenticate with FAS, as `fedora.client.fas2` doesn't support
kerberos.

Resolves #220.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>