Install Docker CE
We started to have this issue with `oc cluster up` yesterday, where the
OpenShift pods would fail to start, so all the tests turned red.

After much time spent and failing to find out the root cause, I've
decided to go with the solution I know it works, that is using Docker
CE, instead of docker from Fedora 30.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>