Make the coverage report configurable in 'make check'
Add a `COV_REPORT` variable defaulting to the current `term-missing`.

This makes it possible to disable the coverage report by setting
`COV_REPORT=""`. This is useful when running the tests locally while
hacking and one is not really interested in the report but rather in the
green or red line at the end of the test run.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>