Report status through PR flags in Pagure
Commit flags are not visible on the PR page in Pagure, and commit flags
and PR flags are separate things in these forges.

This behaviour diverges from how other forges treat and display commit
flags (aka commit checks or commit status): display the flags of the
PR's head commit on the PR page.

In order to replicate this behaviour in the case of Pagure, PR flags
need to be updated when updating commit flags. This should only happen
when the commit is part of a pull-request AND it's the head commit of
that pull request.

Relates to packit-service/ogr#381.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>