Override package_config from job.overrides
Before running a handler for a particular job, override package_config
attributes from job.overrides.

This will enable people to override things per job and do really crazy
stuff, most notably matrices - set up different configurations for
builds, releases, etc.

The reason this is so big is that we were not processing correctly all
jobs with the same type, which will be easy to do with overrides and
multiple triggers. Hence this makes sure that *all* matching jobs are
executed properly, not just a first one.

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>