Restructure the log pages
Log pages for Copr builds had a lazy structure, often causing confusion
for our users. There were cases when it was not obvious for the user if
the SRPM or the Copr build failed.

Improve this by linking SRPM build logs instead of including them. Also
use Flask templating instead of returning raw HTML and add a touch of a
style.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>