new test case: make sure spec is discovered and set
Needs https://github.com/packit-service/packit/pull/886

This is just a p-s' side of #714 in terms of testing so we're sure this
is caught both ways. Once #886 is merged, this should start passing.

TL;DR, make sure that when specfile_path is not set in packit.yaml, it's
discovered correctly and set correctly in all jobs

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>