Turn PR labels into a set in the handler
PR #792, made PullRequestLabelPagureEvents JSON serializable, but this
also means that the handler receives these events as a list, not as a
set.

Due to this the intersection used to find out whether the service should
react on the labels applied fails with an AttributeError ('list' object
has no attribute 'intersection'). To fix this, explicitly construct a
set in the handler's `__init__()`.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>