[Dockerfile(s)] use WORKDIR properly
From https://docs.docker.com/engine/reference/builder/#workdir

"The WORKDIR instruction sets the working directory
for any RUN, CMD, ENTRYPOINT, COPY and ADD instructions
that follow it in the Dockerfile."

So having the WORKDIR instruction at the end of a Dockerfile
(as we had in Dockerfile.worker) makes no sense.

With having it at the beginning we can remove all the 'cd /src/'.

This commit also removes ENVs defined in base image.