Do not post on pull requests if cannot set statuses
We were posting MR comment in case we do not have permissions to set commit
status, we did that only on GitLab, which also proved the code to be harder
to understand. Switching to only commit comments since it makes it more
clear and the behaviour is unified between forges.

Signed-off-by: Matej Focko <mfocko@redhat.com>