Allow webhook token for whole group/namespace
With Group Hooks (Gitlab EE only) we can setup a webhook for
whole group/namespace.
In that case the same token would be sent with events
from each repo, so the token can't be encoded with a repo name,
only with the namespace.

When we receive/decode the token we now accept tokens
without repo_name as well.