Make the MR targets handled in GitLab configurable
'gitlab_mr_targets_handled' is an optional configuration field for Packit
Service, which if defined, is a list of objects with a 'repo' and
'branch' fields. Both of these should be regex strings.

The regex strings of 'repo' are matched against the full slug of the
target repo of the MR ('redhat/centos-stream/src/make', for example).

The regex strings of 'branch' are matched against the name of the target
branch of the MR.

If any of 'repo' or 'branch' is missing, it's considered to match all
repositories or branches.

If 'gitlab_mr_targets_handled' is not specified, all target repositories
and branches are considered to be handled.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>