Fix checking the status of Testing Farm pipelines
Packit Service implementation mixes the pipeline state and result values
into a single enum, called TestingFarmResult. This lead to a
misunderstanding when determining the state of pipelines in the
baby-sitting task.

Fix this by first parsing the response received from Testing Farm, and
then looking at the value of 'result' to tell the state.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>