Remove name of the app/service from check names
The check run name is already quite long and I need to add even more info (test distro) to it.
I checked some checks of various apps and most of them doesn't contain name of the app, so I think it's not needed.
We have different icon for staging app and in the Checks tab the checks are in sections with the app names.

Also the job and target are now separated with ':' instead of '-', because there already are more '-'
in most target names and this way it's more readable.