Discard old SRPM build logs and artifact urls
We're doing this because the SRPM build logs take the most space
in the database and we don't need the old logs.

Periodically (daily at 1AM) check SRPM builds in database and
discard logs and urls to artifacts (uploaded .src.rpm files) at copr
of the builds which are older than SRPMBUILDS_OUTDATED_AFTER_DAYS.
The SRPMBUILDS_OUTDATED_AFTER_DAYS can be set also as env. var.

The artifact urls could be removed even sooner
because copr deletes the artifacts after 2 weeks.
We could also do it similar way copr does, i.e.
leave the latest (for each PR) srpm (build log, url)
indefinitely and discard older builds.
But we might soon be building srpms in copr
so it's probably not worth the effort at the moment.