Make a build/test run naming more explicit
Since there is a plan to have a new model for a group of builds or test
runs, it is better to be explicit using the "Target" word to emphasize
that it is a single run. It has been decided to use this in favour of
the word "Chroot", since the word could be confusing and doesn't apply
well to Testing Farm and for consistency, TF, Copr and Koji should all
use the same naming rules.

Related to this change is the naming of grouped builds, where being
explicit is also recommended, i.e. using the word "Group" inside the
model name.

Signed-off-by: František Nečas <fnecas@redhat.com>