Make the path to the service config configurable
Before this, the configuration file needed to be created under the home
directory of the user running the service.

But during the transition to use mod_wsgi-express (instead of httpd) to
run the service I hit an issue, where 'Path.home()' resolved to the
current working directory when the code was running through
mod_wsgi-express, even though just before starting the service it was
correctly resolving to the home directory of the 'packit' user.

Unfortunately I wasn't able to figure out why this was happening. One
thing I've tried was to set the '--user' option of mod_wsgi-express to
'packit'. But this resulted in an 'invalid user' error.

To work around this, I introduced the possibility to make the path to
the configuration file configurable through a PACKIT_SERVICE_CONFIG
environment variable.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>