Use 'mod_wsgi-express' instead of 'httpd'
This does a few things:

1. Removes all httpd configuration. The web-server is configured by
   mod_wsgi-express, and we use mod_wsgi-express command line option to
   tweak things.
2. Removes some unnecessary dependencies.
3. Adds exiting with an error if the migration before starting the
   web-server fails for any reason. This wasn't happening before this,
   the web-server just started even if the migration failed.
4. Removes running 'setup_env_in_openshift' before the web-server. This
   doesn't seem to do anything at this point.

'--process' is set to 2, b/c this seems to be producing the same memory
footprint as the previous configuration (which it doesn't seem to be
possible to replicate with the command line options, but maybe it's not
necessary).

'--locale' is configured in order to avoid warnings at startup regarding
not being able to set the locale to en_US.UTF-8.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>