compose: Allow the server name to be fully configurable
Don't require the service to bind to 'dev.packit.dev' when running
things with docker-compose, but instead allow the server name to be
fully specified, so that it's possible to configure something like
'service.localhost'.

Additionally: take the server name and the HTTPS port numbers provided
as options for mod_wsgi-express from 'packit-service.yaml'.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>