Report better message when we fail updating our chroot set
The original problem is a limitation in Copr that you cannot update
chroots of a project while builds are running.

Unfortunately, Packit did not detect this well and reported confusing
message to users about permissions:

https://github.com/packit/hello-world/pull/677#issuecomment-1065106598

This commit fixes the particular case and provides a more helpful error
message to users how to proceed and what the problem is.

On top of it, if project chroots require update, the message now
contains a markdown-formatted diff.

Related: https://github.com/packit/packit/issues/1303

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>