Run jobs only configured for the current deployment
Using the packit_instances config key,
user can specify the instances that will accept this job.

By default, just production is accepted,
so we don't break anything for our prod users.

The only change will be required for our stage users
that need to set this key
and define the services that will handle their jobs.

Signed-off-by: Frantisek Lachman <flachman@redhat.com>