Remove JobMetadataConfig
Fix packit#1515

This commit removes also the support for old osolete key dist-git-brach

Usage of metadata key in the yaml files is still supported.
If the same key, let say branch, is used both netsted to job and
netsted to job -> metadata then the job -> metadata -> branch value
is used inside the code.