Fix a mocked call of 'get_aliases()'
When creating an expectation for a function call with the help of
flexmock, the module mocked should be the one in which the function is
imported and used, in order to be properly replaced.

In this particular example although 'get_aliases()' is _defined_ in
'packit.config.aliases', when used, it's imported in
'packit_service.worker.helpers.build.copr_build' with

    from packit.config.aliases import get_aliases

and so it becomes a name in 'copr_build'. This is why it's 'copr_build'
that needs to be mocked, instead of 'packit.config.aliases'.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>