Don't add duplicated commit comments
Fixes https://github.com/packit/hardly/issues/29
Requires https://github.com/packit/ogr/pull/717

The commit comments are added only as a fallback when adding a commit status fails
(because of permissions or forge issues).
It can happen that we want the set a status that a commit already has.
With commit statuses it's not a problem, but in the fallback mode
when we add comments instead, it results in duplicated comments.