Avoid duplicating ProposeDownstreamTargetModel when retrying the task
When retrying the propose downstream task, we pass the ID of ProposeDownstreamModel so that
it is created only once, but for each retry we created new ProposeDownstreamTargetModel instances
which in combination with zipping with branches created incorrect records in database (incorrect statuses,
mismatch between model branch and the actual branch, running propose downstream for a branch that already
succeeded). With these changes, for each branch there is created ProposeDownstreamTargetModel only once.