Correctly report results from Testing Farm
Previously, if there were multiple TF jobs differentiated by their
identifier, Steve would create a celery task for each of the jobs,
regardless of their identifier for the single results event. To overcome
this, save the identifier in the DB and obtain it when a result arrives.
Check for identifier match when creating celery tasks based on jobs.

Signed-off-by: František Nečas <fnecas@redhat.com>