Don't handle installation 'added' events
These are the events when repositories are being added to the app installation,
but we were not updating the 'repositories' in the DB until the changes I made in #1575,
which were meant for handling reinstallation correctly (to have the new sender_login and created_at data),
so 'repositories' is only rewritten by the added repos which is not correct.
Also, we don't handle the events when repositories are being removed from the app installation,
so with the current codebase, it doesn't make sense to react to these.