Use singleton SQLAlchemy session for workers
The Contextual/Thread-local session registry, as created by scoped_session(),
is needed for httpd, which uses the traditional threads.
https://docs.sqlalchemy.org/en/14/orm/contextual.html

It's however not compatible with alternative forms of concurrency control,
i.e. greenlets as provided by the eventlet/gevent libraries,
which celery uses for alternative execution pool implementation.
Running multi-(green)threaded worker with it results in
https://docs.sqlalchemy.org/en/14/errors.html#queuepool-limit-of-size-x-overflow-y-reached-connection-timed-out-timeout-z

Using a single session, kept as a global variable, is the only way I've
found that doesn't throw bunch of other errors (like 'Instance not bound
to a session' or 'Object already attached to a session') and doesn't
look more like a workaround rather than a clean code.
We actually used it like this before 631a4a9f3 so it should be fine per se.

Fixes #1591