Update checking Testing Farm composes
The endpoints for getting the TF composes were updated: 'https://api.dev.testing-farm.io/v0.1/composes/redhat' for
internal TF and 'https://api.dev.testing-farm.io/v0.1/composes/public' for the public. Update these in the code
and check the existence of compose also for internal TF. Implement also the reporting if the distro
cannot be mapped to a compose.