Keep track of copr build used for TF run
After introduction of groups, there is no longer a 1:1 mapping between
runs and copr builds (there is 1:1 between run and group and 1:N between
group and targets), hence it is no longer possible to simply obtain copr
build. For this reason, we need to keep track of the corresponding build
and obtain it later.

Signed-off-by: František Nečas <fnecas@redhat.com>