Use --concurrency instead of --autoscale
The --autoscale sets concurrency (number of threads) dynamically based on load (number of tasks in a queue).
In theory, this sounds like it should save some resources when not needed
while still being able to handle lots of requests during peaks.
In practice, I don't see any difference in used resources (mem,cpu) when I set --autoscale=4,4 or --autoscale=100,100.
i.e. it still uses the same amount of resources regardles whether there are 4 threads or 100 threads.
So, at the moment, I don't see a use-case for the --autoscale and plain --concurrency should be enough
(easier to understand and no overhead in scaling the pool of threads).