Implement triggering TF with build from other PR
Implement getting of additional Copr build target models that can be specified
by test comment command PR argument, e.g. '/packit test repo/namespace#123'.
In TestingFarmJobHandler pre-check check whether there are any successful recent Copr builds
for that PR in the database and report to user.

If TF is triggered by a comment with PR argument, include the artifact
for the build from that PR, the packages from that build in the payload
and add the TF run to the pipeline for that PR build in the database as well.
Change the API for getting the TF runs to be able to get
the additional build from the TF run.

Also split the long methods in the TestingFarmJobHelper.