Move the check for event and job identifiers equality
Move the check whether identifier of the event we are reacting to
 is equal to the identifier of the currently processed job
to get_jobs_matching_event method. Previously, the check was in the
get_handlers_for_event method  which calls the get_jobs_matching_event. This was
not enough, because get_jobs_matching_event is being called 2 times, once more
after running get_handlers_for_event and therefore handler was triggered
with more job configs than correct.

Fixes #1625