testing farm composes: pick distro if user requested it
This is a follow up to c0ca4cd18758f6cd1de36607978b90271b08950d

The problem is that we suffix target names with {arch} and hence why we
need this code to make the use case above work.

Also added more `.debug()` lines in our code so we can spot errors like
these better in future. I could not find `target={target}` log lines in
sentry so it was hard to deduce the content of the variable.

Related #1653

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>