Remove _test_check_names attribute from BaseBuildJobHelper
For BaseBuildJobHelper, methods reporting the testing farm job
statuses work with job configs passed
in method arguments, therefore it doesn't make sense to store
the results in the attribute and reuse it. Move this to TestingFarmJobHelper
where we handle always only one test job and therefore it makes
sense to store it there.