Do not bump with empty release_suffix
Do bump means do update version.
The bump flag wins over the release_suffix value,
this is obvious when using the command line tool
but less natural when using packit-service.
Since the code is the same for both CLI and service
we need to force the bump flag, calculated by the service,
 to be False when release suffix is empty.

Co-authored-by: Matej Focko <mfocko@users.noreply.github.com>