Move monitoring.py:measure_time() to utils.py:elapsed_seconds()
so it's together with
is_timezone_naive_datetime() and get_timezone_aware_datetime()

- Change get_timezone_aware_datetime() so that it calls
is_timezone_naive_datetime() first.
- Changed some datetime.utcnow() to datetime.now(timezone.utc)

Initially, I just wanted to fix the Sourcery warnings
https://docs.sourcery.ai/Reference/Built-in-Rules/suggestions/aware-datetime-for-utc
(to create timezone-aware datetime objects instead of timezone-naive)
In the database, we however still store the values timezone-naive,
so we have to use utcnow() when setting those values.

We already have the above functions (added in #1251)
for safe comparison/subtraction of timezone-aware and timezone-naive
values (to avoid TypeError), but the ultimate solution would be
to change the datetimes in the database to be timezone-aware. (#1749)