Return only test job from job_tests_all if working with test job config
If working with test job config (in CoprBuildJobHelper), we don't want
to utilise all test job configs (there is no "central" job configuration
and for each test job config the build is done separately).
We want to do this only if working with build job config (=> build job is defined).