Simplify the actor check for internal TF
Previously, we needed to check for internal TF usafe even when
running CoprBuildHandler, because we did not store the 'actor'
to be checked and therefore the check needed to be run when processing the
webhook event.
This commit adds the 'actor' field to the PullRequestModel, so we can
run the actor check only when running the Testing Farm directly.

Fixes #1725