Export variables set in run_worker.sh
Initially they were used only in the same script to pass arguments to 'celery worker' command.
But now we use them also in models.py to see whether we're running multi-threaded.
This change is needed when we don't explicitly set POOL as pod's env. variable, but set
it in the script based on the CONCURRENCY value (5eb8c83743).