Implement PullFromUpstreamHandler
The handler reacts to NewHotnessUpdateEvent and uses the functionality
of ProposeDownstreamHandler. There is also a new helper class for it and both
helper and handler classes have the common functionality in the parent classes.
For the new handler there is a pre_check method to report to user if
the project_url or upstream_tag_template was misconfigured via issue_repository.
TBD: changes regarding storing the info to DB