SLO1: raise limit from 15s to 30s
If a job defines a lot of targets (10+), we can easily run out of SLO1
because it takes 1s to set one check. With the current implementation,
we measure how long it took to set every single status line - even the
last one.

We want to preserve this to be sure all are set in a reasonable time.

We have picked 30s as the next step & conservative value. If this
constant does not work, we should revisit.

Precise description:
https://github.com/packit/packit-service/issues/1728#issuecomment-1326534470

Fixes #1728

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>