fix(vm-image-build): check Copr projects match
build.project_name is a name of a project in Copr, we need to match
owner/project of a build with those specified in the vm-image-build
owner/project to make sure that the latest build for the commit is done
and successful.

We cannot match binary RPMs since we don't know what exactly is user
trying to install in the target image.

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>