change(celery): retry backoff 3 → 7
This also changes retry setup for creation of bodhi updates to 5.
All other tasks are still on 2.

When will the retrying happen?

  0-7s, 0-14s, 0-28s, 0-48s, 0-96s

We have jitter on (celery's default) so it's actual `random(0, value)`.

This is done mainly to account for koji build tagging as it can take
tenths of seconds easily. We also wanted to increase the backoff since
it makes more sense to retry later: the initial condition may not be
resolved right away.

Brutally related #1615

Signed-off-by: Tomas Tomecek <ttomecek@redhat.com>
Co-authored-by: Laura Barcziova <lbarcziova@redhat.com>