Create new handlers for re-triggering through an issue
The koji builds, bodhi updates, propose-downstream jobs
can be retriggered through comments in a issue (in the
issues repository) and the issue repository does not
need the packit yaml.

When re-triggering a bodhi update or koji build through
a comment in a issue (in the issues repository) do not use
local_project or project_url from event.
They reference the issue repository and not the distgit repo.
Packit API have to be built using the distgit url
found in the issue description and saved in a new property
inside the AbstractIssueCommentEvent.

All the failed branches (written inside the issue description)
will be retriggered by a single re-trigger comment.