COPY only necessary files
to avoid reinstalling all dependencies if one only changes something in this Dockerfile in later steps.
This Dockerfile is one of the files/* and if I change something even at the very end of the file
everything was rebuilt since the 'COPY files/' step, i.e. including installing all dependencies again.