Reporting handlers improvemets for idempotency
See https://github.com/packit/packit-service/issues/1692#issuecomment-1386664671,
for the handlers that are reporting state change, in the beginning always check
whether the state in DB is different that the state that we are about to report
and change the state in the end of the handler code.
Also improve other operations, e.g. creating issue in notifications repo
will now check for existing duplicate issues.