Get the actual latest Koji build when retriggering Bodhi update
Koji tags named equally to dist-git branches are what stable builds
are tagged into, but when creating a Bodhi update, we are interested
in the latest non-stable build that is tagged into a candidate tag.

Get candidate tag corresponding to the dist-git branch being processed
and use that to get the latest build from Koji.

Signed-off-by: Nikola Forró <nforro@redhat.com>