fix(bodhi): auth failure is not expected anymore
After we were given the global rights to run Bodhi updates, we don't need
to ask for ‹commit› privileges on dist-git, therefore this kind of failure
is now unexpected and should be reported to Sentry.

Also we shouldn't suggest adding ‹commit› access to the dist-git as it is
no longer needed.

Related to #1721

Signed-off-by: Matej Focko <mfocko@redhat.com>
Co-authored-by: Laura Barcziová <lbarczio@redhat.com>