Solve broken monorepo test using get_job_views
In this way we will not touch our handlers but we are forced to
handle one package at time.

If we need to deal with multiple packages at the same time
we need to create new handlers and a new way to get them
from a new call, something like the
package_config.get_grouped_job_views().
We will pass to the new handlers a list of ConfigJobView(s)
instead of a single [ConfigJob|ConfigJobView]