fix: do not request access to GitLab forks
When creating commit statuses, we need to have access to the repository
having the commits, in case of one-time contributors automatic requesting
of access can be an unnecessary noise and spam.

Do not request the access and just post a note about the possibility to
add the GitLab user to the said repository.

Fixes #1892

Signed-off-by: Matej Focko <mfocko@redhat.com>