Update parsing of Gitlab [tag] push events
There was a problem that as a head commit we selected the last item
from the list of commits (head_commit = commits[-1]),
but it seems that it's actually the first item in the list.
Rather than assuming which one it is, select it based on already known sha of the head commit.

Also get_gitlab_push_common_data() does not need to
return head_commit_sha as it already is in the head_commit.

The rest is just shuffling code and some walruses.