Unify checking of the trigger branch name in build and test handlers
When the trigger type of the configured job is a commit, for builds and tests
we additionally need to check whether the branch name from the trigger matches.
Before the change, this was done only in Copr and Koji build handler pre-checks:
unify the method that checks the match and use it also in TF handler pre-check.
Fixes #1981