test: add test case to check merging of TF configs
In the end, it was a typo on the config-side. Environments are taken by
the API as an array (for easier customizability), in this case it was
supposed to be one object rather than two »separate« objects in the config.

Adding a test case to »explicitly« cover this case in case of regressions.

Fixes #2010

Signed-off-by: Matej Focko <mfocko@redhat.com>