Sentry: Set-up Transactions/Performance
Setting the traces sampler seems to be all we need to see the Transactions/Performance.

https://docs.sentry.io/platforms/python/performance
https://docs.sentry.io/platforms/python/performance/instrumentation/automatic-instrumentation

From docs:
Choose a sampling rate with the goal of finding a balance between performance and volume concerns with data accuracy.
You don't want to collect too much data, but you want to collect sufficient data from which to draw meaningful conclusions.
If you’re not sure what rate to choose, start with a low value and gradually increase it as you learn more about your traffic patterns and volume.

https://docs.sentry.io/platforms/python/configuration/sampling/#sampling-transaction-events