Add 'complete' to TestingFarmResult enum
The 'state' from TF response can be  "new" "queued" "running" "complete" "error"
(see https://testing-farm.gitlab.io/api/#operation/requestsGet)
and since we parse it in case the 'result' is null, all these states should be in the
TestingFarmResult enum. In general it should not happen that the 'state' is 'complete',
the 'result' is null (but TF can produce incorrect responses sometimes).

Fixes #2038