GitHub: Consider only the events we are interested in
This extends the interested() method in the GitHub webhook, so that
checks on the event type and action are done in the API. This will help
avoid creating Celery tasks for these events.

This change makes similar checks in Parser() redundant, but I've decided
not to remove those, yet, mainly b/c it's hard to tell how that will
influence parsing and telling which event is handled.

Signed-off-by: Hunor Csomortáni <csomh@redhat.com>