Improve mapping the TF state and result to our states
Discussed this with mvadkert over chat:
error and complete are end states fro request
overall.result is for result of testing, passed or failed, in case of error, it cannot be determined.

This should fix reporting in a situation that pipeline errors out,
so the result["overall"] may be "unknown" (which we would report
as failure before) but TF does the autoretrying and the state is
therefore not yet in the end state, so we should rather report the state.

Related to #2216