Adjust Koji event and model and related methods
These changes are needed for better integration of the implementation
of storing downstream Koji builds in the DB.
Rename build_id to task_id attribute and related methods and make its usage consistent.
Introduce dict attribute of the event rpm_build_task_ids that is a dict of {arch: id}.
Also adjust the method constructing the logs URL to utilise this dictionary.

Fixes #2243
Related to #1889