Do not show build status in manual test checks
This can be improved having a real dependency
between tests and builds.

Showing a neutral state for a successful build
message is not working on gitlab and even on
github is not always working
(my test falls back on commit statuses and
they don't have a neutral state).

I can not show only errors related with build
statuses because they will be not cleaned up by
successful statuses later and this is misleading.

We can just hide them all, at least until we
have real dependencies.