fix(sync-release): upload to lookaside cache only for Fedora
When uploading the sources to lookaside cache, we cannot upload them to
the CentOS as the cache is writeable only on the VPN / internal network.

However we still want to be able to submit updates via an MR, even if
maintainer has to submit the archives to the lookaside cache by
themselves.

Once the workflow with regards to the lookaside cache has been improved,
drop this requirement from our side.

Signed-off-by: Matej Focko <mfocko@redhat.com>