Don't trigger builds from TestingFarmHandler
This changes to behaviour of /packit test: in case there is missing
build for some target, the build will not be triggered anymore, it will
just be reported to the user.

Previous code was prone to the race condition if there were multiple test jobs in configs,
as there were scheduled multiple TestingFarmHandler runs where
each could trigger Copr build jobs. This potentially meant wasting of lot of resources.