Run job_config handlers in new tasks
If we run sequentially all the handlers for all
the job_configs, for every entry in the db
that needs babysitting we can reach
the hard time limit for the task.

I did not make the vm image build handlers
running in parallel, because now the code
needs the tasks output.

Co-authored-by: Nikola Forró <nforro@redhat.com>