feat(events): implement ‹event_type()›
Till now we have used ‹.__name__› on the event classes for serialization
for the Celery, but after the refactor the ‹.__name__› contains just the
last name in the whole path (i.e., only the class name is included,
modules are lost) which creates conflicts and also breaks the matching
of serialized ‹event_type› to the respective classes.

Therefore implement a new abstract class method that returns the whole
event type (including the source and any additional subtyping).

Also implement the ‹event_type()› for some of the abstract classes that
are used in the tests, with a check that the tests are running to ensure
that abstract classes are not constructed during the runtime.

Signed-off-by: Matej Focko <mfocko@redhat.com>