fix(parser): refactor common parsers to data classes
When we started sharing parsers for parts of webhooks that have same
structure, we used to return tuples, but in the current state it leads to
returning 5 or even 10 members in such tuple. This is hard to keep track
of and not very extensible in the future.

Therefore switch to return data classes instead.

Signed-off-by: Matej Focko <mfocko@redhat.com>