fix(db): do not yield running IDs directly
As part of the follow-ups for cancelling the jobs, it is also necessary
to update the statuses of the builds. Yielding just the IDs that are to
be cancelled complicates this as it would require going through the DB
entries again to find the ones that are cancelled to update the status.

Therefore yield directly the target models, so it is possible to do some
filtering or such on the “user-end” rather than directly in ‹models.py›.

There are also some inefficiences with regards to the relationship
between Copr build and Copr targets, this is to be addressed separately
as it involves more aggressive changes…

Signed-off-by: Matej Focko <mfocko@packit.dev>