fix: do not log failed submodule reset as an error
Do not log failure when resetting the git submodule during sync-release
as an error.

Since it gets logged as an error, Sentry takes it as is, and also for
each unique submodule, tracks as a separate error.

This should not be an issue, as we have already discussed with @nforro,
therefore it should be enough to be logged as warning instead.

Signed-off-by: Matej Focko <mfocko@packit.dev>