fix: correct the type after introducing TF Client
Another thing missed from the factoring out of the Testing Farm client.
It is now a separate property and not under the helper.

Additionally also shorten the name of the method used for canceling the
requests. No need to be so explicit, as long as it's TF client, it's
apparent we are canceling TF requests.

Fixes #2793

Signed-off-by: Matej Focko <mfocko@packit.dev>