fix(tf-client): introduce `default_ranch` property
Since most of the actions done by the TF Client are tied to the specific
job config, we pass `use_internal_tf` value into the client itself.

However client doesn't have to be tied to the config itself. Since there
is still some code tied specifically to the redhat ranch, I cannot simply
drop the `use_internal_tf` in favor of passing the ranch directly to the
constructor.

We should probably think about it a bit more…

Signed-off-by: Matej Focko <mfocko@packit.dev>