fix(tf): differentiate ranches when canceling
As we are given different API tokens for `public` and `redhat` ranch, it
is not possible to cancel the running tests on public ranch using redhat
token and vice versa, therefore we need to be able to differentiate
between the ranches where the tests are being run.

Since we can trigger multiple targets on one ranch and they are all tied
to one test job in Packit config, it is possible to keep `ranch` on the
TF test group model.

It is also possible to deduce the ranch of the previous tests based on
the URL (URL is kept on the target; Packit test target = TF pipeline).

When canceling Testing Farm runs, filter additionally by the ranch to not
fail on the 403.

Fixes #2808

Signed-off-by: Matej Focko <mfocko@packit.dev>